adventure/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-ai-ml==1.17.0
azure-core==1.31.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distro==1.9.0
docstring_parser==0.16
fire==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
isodate==0.6.1
jiter==0.5.0
mistral_common==1.3.4
mistral_inference==1.3.1
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.45.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
termcolor==2.4.0
tiktoken==0.7.0
torch==2.4.0
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
xformers==0.0.27.post2
